+++ src/nationalarchives/components/cookie-banner/cookie-banner.stories.js
@@ -1,6 +1,6 @@
 import CookieBanner from "./template.njk";
 import macroOptions from "./macro-options.json";
-import { expect } from "@storybook/jest";
+import { expect, jest } from "@storybook/jest";
 import { within, userEvent } from "@storybook/testing-library";
 import Cookies from "../../lib/cookies.mjs";
 
@@ -8,7 +8,6 @@ const argTypes = {
   cookiesUrl: { control: "text" },
   policies: { control: "text" },
   cookiesPreferencesSetKey: { control: "text" },
-  loadScriptsOnAccept: { control: "text" },
   classes: { control: "text" },
   attributes: { control: "object" },
 };
@@ -28,7 +27,6 @@ const Template = ({
   cookiesUrl,
   policies,
   cookiesPreferencesSetKey,
-  loadScriptsOnAccept,
   classes,
   attributes,
 }) =>
@@ -37,7 +35,6 @@ const Template = ({
       cookiesUrl,
       policies,
       cookiesPreferencesSetKey,
-      loadScriptsOnAccept,
       classes,
       attributes,
     },
@@ -49,8 +46,15 @@ Standard.args = {
   classes: "tna-cookie-banner--demo",
 };
 
-const deleteAllCookies = (cookies) => {
-  Object.keys(cookies.all).forEach((cookie) => cookies.delete(cookie));
+const deleteAllCookies = () => {
+  const allCookies = {};
+  document.cookie.split(";").forEach((cookie) => {
+    const parts = cookie.trim().split("=");
+    allCookies[parts[0]] = parts[1];
+  });
+  Object.keys(allCookies).forEach((key) => {
+    document.cookie = `${encodeURIComponent(key)}=`;
+  });
 };
 
 export const Accept = Template.bind({});
@@ -59,8 +63,9 @@ Accept.args = {
   classes: "tna-cookie-banner--demo",
 };
 Accept.play = async ({ canvasElement }) => {
+  deleteAllCookies();
+
   const cookies = new Cookies();
-  deleteAllCookies(cookies);
 
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
@@ -87,7 +92,7 @@ Accept.play = async ({ canvasElement }) => {
 
   // await expect(closeButton).not.toBeVisible();
 
-  deleteAllCookies(cookies);
+  deleteAllCookies();
 };
 
 export const Reject = Template.bind({});
@@ -96,8 +101,9 @@ Reject.args = {
   classes: "tna-cookie-banner--demo",
 };
 Reject.play = async ({ canvasElement }) => {
+  deleteAllCookies();
+
   const cookies = new Cookies();
-  deleteAllCookies(cookies);
 
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
@@ -118,7 +124,7 @@ Reject.play = async ({ canvasElement }) => {
   await expect(acceptButton).not.toBeVisible();
   await expect(rejectButton).not.toBeVisible();
 
-  deleteAllCookies(cookies);
+  deleteAllCookies();
 };
 
 export const CustomPolicies = Template.bind({});
@@ -128,8 +134,9 @@ CustomPolicies.args = {
   classes: "tna-cookie-banner--demo",
 };
 CustomPolicies.play = async ({ args, canvasElement }) => {
+  deleteAllCookies();
+
   const cookies = new Cookies(args.policies.split(","));
-  deleteAllCookies(cookies);
 
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(null);
@@ -145,33 +152,33 @@ CustomPolicies.play = async ({ args, canvasElement }) => {
   await expect(cookies.isPolicyAccepted("settings")).toEqual(null);
   await expect(cookies.isPolicyAccepted("custom")).toEqual(true);
 
-  deleteAllCookies(cookies);
+  deleteAllCookies();
 };
 
-export const AddScriptsOnAccept = Template.bind({});
-AddScriptsOnAccept.args = {
-  cookiesUrl: "#",
-  loadScriptsOnAccept: "my-usage-script.js",
-  classes: "tna-cookie-banner--demo",
-};
-AddScriptsOnAccept.play = async ({ args, canvasElement }) => {
-  const cookies = new Cookies();
-  deleteAllCookies(cookies);
+// export const EventHandling = Template.bind({});
+// EventHandling.args = {
+//   cookiesUrl: "#",
+//   policies: "custom",
+//   classes: "tna-cookie-banner--demo",
+// };
+// EventHandling.play = async ({ args, canvasElement }) => {
+//   deleteAllCookies();
 
-  const noScript = document.querySelector(
-    `script[src="${args.loadScriptsOnAccept}"]`,
-  );
-  await expect(noScript).toEqual(null);
+//   const cookies = new Cookies();
 
-  const canvas = within(canvasElement);
-  const acceptButton = canvas.getByText("Accept cookies");
-  await userEvent.click(acceptButton);
+//   const onChangePolicy = jest.fn(data => {
+//     console.log(data)
+//   })
+//   cookies.on("changePolicy", onChangePolicy)
 
-  const script = document.querySelector(
-    `script[src="${args.loadScriptsOnAccept}"]`,
-  );
-  await expect(script).toBeTruthy();
+//   const canvas = within(canvasElement);
+//   const acceptButton = canvas.getByText("Accept cookies");
+//   await userEvent.click(acceptButton);
 
-  deleteAllCookies(cookies);
-  script.remove();
-};
+//   await expect(onChangePolicy.mock).toHaveBeenCalledTimes(1);
+//   await expect(onChangePolicy.mock.calls).toHaveLength(1);
+//   await expect(onChangePolicy.mock.results[0].value).toHaveProperty("custom");
+//   await expect(onChangePolicy.mock.results[0].value.custom).toEqual(true);
+
+//   deleteAllCookies();
+// };