+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Grid columns can be aligned along the cross axis with classes `tna-column--align-top`, `tna-column--align-bottom` and `tna-column--align-centre`
 - Actions and items using the `href` attribute on cards can now have custom classes and attrbiutes added
+- Events can now be added when cookie preferences change
 
 ### Changed
 
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed CSS to counter conflicting GOV.UK paragraph styling
 - Transitions removed from most elements
+- `loadScriptsOnAccept` option for cookie banner removed in favour of callback events
 
 ### Fixed
 
@@ -31,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Colour fixed for plain card supertitles in accented style
 - Header navigation uses flexbox for layout of tabs and top navigation stacks on mobile
 - Explicitly declare font colour for tinted and light accent backgrounds
+- Accepting or declining individual cookie policies now works
 
 ### Security